add recruiter profile and recruiter APIs

+++ database/migrations/2025_02_22_040911_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('applicant_id');
             $table->unsignedBigInteger('company_id');
-            $table->double('rating')->nullable();
+            $table->float('rating')->nullable();
             $table->text('comment')->nullable();
             $table->integer('lock_version')->default(0);
             $table->timestamps();
